requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Cython==3.0.12
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
emoji==2.14.1
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kombu==5.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psycopg2-binary==2.9.10
PyAudio==0.2.14
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
razorpay==1.4.2
redis==5.2.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
SpeechRecognition==3.14.1
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
stanza==1.10.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
pytesseract==0.3.10
transformers==4.38.2
prophet==1.1.5
xgboost==2.0.3
